deleted sort for view groups

+++ app/Http/Controllers/GroupController.php
@@ -5,7 +5,6 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Models\City;
 use App\Models\Group;
-use App\Models\User;
 use Illuminate\Support\Facades\Auth;
 
 class GroupController extends Controller
@@ -35,7 +34,6 @@ class GroupController extends Controller
         if (empty($group)) {
             return redirect()->route('groups.index')->with('alert', 'Группа не найдена');
         }
-
 
         $subscribe = false;
 
@@ -87,12 +85,19 @@ class GroupController extends Controller
     }
 
     public function unsubscribe($id)
-    {
-        
+    {      
         $group = Group::find($id);
 
         if (empty($group)) {
             return redirect()->route('groups.index')->with('alert', 'Группа не найдена');
+        }
+
+        if (!Auth::check()) {
+            return redirect()->route('group.show', ['id' => $group->id])->with('alert', 'Необходимо войти');
+        }
+
+        if (!$group->isOfThe(Auth::user())) {
+            return redirect()->route('group.show', ['id' => $group->id])->with('alert', 'Вы уже подписаны');
         }
 
         $group->users()->detach(Auth::user()->id);
